move most urls/views to frontend app
